Move generate_data_fields utility to services app
This utility manages models from the services app, so it should reside
in the same app.
