+++ utils/utils.py
@@ -12,23 +12,6 @@ from profiles.models import Address, Email, Phone, Profile
 from services.models import AllowedDataField, Service, ServiceConnection
 from users.models import User
 
-
-@transaction.atomic
-def generate_data_fields(allowed_data_fields_spec):
-    """Create data fields if they don't exist."""
-    for value in allowed_data_fields_spec:
-        if not AllowedDataField.objects.filter(
-            field_name=value.get("field_name")
-        ).exists():
-            data_field = AllowedDataField.objects.create(
-                field_name=value.get("field_name")
-            )
-            for translation in value.get("translations"):
-                data_field.set_current_language(translation["code"])
-                data_field.label = translation["label"]
-            data_field.save()
-
-
 SERVICES = [
     {
         "name": "berth",
